backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.10.0
torchvision==0.25.0
pillow
numpy
opencv-python
scikit-image
matplotlib
timm
